+++ test_fsictools.py
@@ -76,6 +76,9 @@ class TestPandasFunctions(unittest.TestCase):
 
         pd.testing.assert_frame_equal(result, expected)
 
+    def test_linker_to_dataframes(self):
+        raise NotImplementedError
+
 
 networkx_installed = True