+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 
 ### Added
 
+- New `BaseLinker` class to nest and solve multiple `BaseModels`.
 - New `BaseModel` class method, `from_dataframe()`, to instantiate a model
   object from a `pandas` DataFrame or similar.
 - Further keyword arguments to `BaseModel.solve()` are now passed to both